Move worker-python preparation logic into the prepared module
config.rs is now a pure structural definition of Config and its
builders. The resolution work — default_runner and the PATH
executable-lookup family (find_executable/is_executable_file/
is_executable) plus their tests — moves next to its only consumer,
Config::prepare in prepared.rs.

Addresses the two organizational review comments on PR #423.
